requirements.txt
airtest==1.1.6
certifi==2020.12.5
chardet==4.0.0
hrpc==1.0.8
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.19.4
opencv-contrib-python==3.4.2.17
Pillow==8.0.1
pocoui==1.0.76
python-xlib==0.29
pywinauto==0.6.3
requests==2.25.1
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
